SASS-5703 content for Royalty payments page
