+++ app/forms/aboutyou/PatentRoyaltyPaymentsFormProvider.scala
@@ -14,13 +14,24 @@
  * limitations under the License.
  */
 
-package pages
+package forms.aboutyou
 
-import play.api.libs.json.JsPath
+import forms.mappings.Mappings
+import play.api.data.Form
 
-case object PatentRoyaltyPaymentsPage extends QuestionPage[Boolean] {
+import javax.inject.Inject
 
-  override def path: JsPath = JsPath \ toString
+class PatentRoyaltyPaymentsFormProvider @Inject() extends Mappings {
 
-  override def toString: String = "patentRoyaltyPayments"
+  def apply(isAgent: Boolean): Form[Boolean] = {
+    val error: String =
+      if (isAgent) {
+        "patentRoyaltyPayments.agent.error.required"
+      } else {
+        "patentRoyaltyPayments.error.required"
+      }
+    Form(
+      "value" -> boolean(error)
+    )
+  }
 }
