+++ app/pages/aboutyou/PatentRoyaltyPaymentsPage.scala
@@ -14,24 +14,14 @@
  * limitations under the License.
  */
 
-package forms
+package pages.aboutyou
 
-import javax.inject.Inject
+import pages.QuestionPage
+import play.api.libs.json.JsPath
 
-import forms.mappings.Mappings
-import play.api.data.Form
+case object PatentRoyaltyPaymentsPage extends QuestionPage[Boolean] {
 
-class PatentRoyaltyPaymentsFormProvider @Inject() extends Mappings {
+  override def path: JsPath = JsPath \ toString
 
-  def apply(isAgent: Boolean): Form[Boolean] = {
-    val error: String =
-      if (isAgent) {
-        "patentRoyaltyPayments.agent.error.required"
-      } else {
-        "patentRoyaltyPayments.error.required"
-      }
-    Form(
-      "value" -> boolean(error)
-    )
-  }
+  override def toString: String = "patentRoyaltyPayments"
 }
