+++ app/views/aboutyou/PatentRoyaltyPaymentsAgentView.scala.html
@@ -26,7 +26,10 @@
 
 @layout(pageTitle = title(form, messages("patentRoyaltyPayments.agent.title")), taxYear = taxYear) {
 
-    @formHelper(action = routes.PatentRoyaltyPaymentsController.onSubmit(mode, taxYear), Symbol("autoComplete") -> "off") {
+    @formHelper(action = controllers.aboutyou.routes.PatentRoyaltyPaymentsController.onSubmit(mode, taxYear), Symbol("autoComplete") -> "off") {
+
+        <h1 class="govuk-heading-l">@messages("patentRoyaltyPayments.agent.title")</h1>
+        <p class="govuk-body">@messages("patentRoyaltyPayments.agent.p1")</p>
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
@@ -35,7 +38,7 @@
         @govukRadios(
             RadiosViewModel.yesNo(
                 field = form("value"),
-                legend = LegendViewModel(messages("patentRoyaltyPayments.agent.heading")).asPageHeading()
+                legend = LegendViewModel(messages(""))
             )
         )
