+++ test/controllers/aboutyou/PatentRoyaltyPaymentsControllerSpec.scala
@@ -14,23 +14,23 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.aboutyou
 
 import base.SpecBase
-import forms.PatentRoyaltyPaymentsFormProvider
-import models.{NormalMode, UserAnswers, Done}
+import controllers.routes
+import forms.aboutyou.PatentRoyaltyPaymentsFormProvider
+import models.{Done, NormalMode, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.PatentRoyaltyPaymentsPage
+import pages.aboutyou.PatentRoyaltyPaymentsPage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import services.UserDataService
-import views.html.PatentRoyaltyPaymentsView
-import views.html.PatentRoyaltyPaymentsAgentView
+import views.html.aboutyou.{PatentRoyaltyPaymentsAgentView, PatentRoyaltyPaymentsView}
 
 import scala.concurrent.Future
 
@@ -42,7 +42,7 @@ class PatentRoyaltyPaymentsControllerSpec extends SpecBase with MockitoSugar {
   val form = formProvider(isAgent = false)
   val agentForm = formProvider(isAgent = true)
 
-  lazy val patentRoyaltyPaymentsRoute = routes.PatentRoyaltyPaymentsController.onPageLoad(NormalMode, taxYear).url
+  lazy val patentRoyaltyPaymentsRoute = controllers.aboutyou.routes.PatentRoyaltyPaymentsController.onPageLoad(NormalMode, taxYear).url
 
   "PatentRoyaltyPayments Controller" - {
